refactor(imports): update prompt module paths to new context structure
- Refactored import statements across multiple files to replace references to the deprecated `agent::prompt` module with the new `context::prompt` module.
- This change enhances code organization and aligns with the recent restructuring of the context management system.
